Protect image files; move passwd and group along; improve suspend.

+++ session-functions.sh
@@ -3,30 +3,26 @@
 export LOCKF=~/.suspend.lck
 
 function activity() {
-	[ -f $LOCKF ] || [ 0$(strafe mlist | wc -l) -gt 0 ]
+	([ -f $LOCKF ] || [ 0$(strafe mlist 2>/dev/null | wc -l) -gt 0 ]) && [ -z "$(pidof xsecurelock)" ]
 }
 
 function sleeper() {
-	SLEEP=$(( 0$1 ))
-	(if [ $SLEEP -gt 0 ] ; then
-		for i in 1 2 3 4 5 6 7 8 9 10 ; do
+	if [ -n "$1" ] ; then
+		for i in 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 ; do
 			if activity ; then
 				rm -f $LOCKF
+				false
 				return
 			fi
-			sleep $SLEEP
+			sleep 1
 		done
-	fi;
-	if activity ; then
-		rm -f $LOCKF
-		return
 	fi
-	yes y | sudo strafe stop-all mprune prune clean
+	(yes y | sudo strafe stop-all mprune prune clean
 	sudo network-stop
 	sudo -k
 	for i in $(cat /proc/acpi/wakeup | grep enabled | cut -f1 -d\ ) ; do echo $i > /proc/acpi/wakeup ; done
-	systemctl suspend 
-	) & disown
+	lxqt-leave --suspend) & disown
+	true
 }
 
 function resumer() {
